refactor: reorganising files into respective folder groups
